+++ UnityProject/Assets/Assets.asmdef
@@ -26,7 +26,8 @@
         "Shared",
         "Mirror.Ignorance",
         "Mirror.Transports",
-        "SecureStuff"
+        "SecureStuff",
+        "SunVox"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
