+++ UnityProject/Assets/Scripts/Core/SecureStuff/VariableViewer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d1173f1af2110ed4ea1269601b677c5c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
