+++ UnityProject/Assets/Scripts/Core/SecureStuff/VariableViewer/ICustomSerialisationSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7dee9dc1aaa26504d8f5935c665b8b5d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
