+++ UnityProject/Assets/Scripts/Core/SecureStuff/VariableViewer/Librarian.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 82fa65fc4030ce646b2ecec96a653fc8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
