+++ UnityProject/Assets/Scripts/Core/SecureStuff/VariableViewer/VVNote.cs.meta
