+++ UnityProject/Assets/Scripts/Core/Sound/SunVox.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e2c34ab0ff368b9498467d761828e528
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
