+++ UnityProject/Assets/Scripts/Core/Sound/SunVox/SunVox.asmdef
@@ -1,0 +1,4 @@
+{
+	"name": "SunVox",
+	"references":[ "GUID:343deaaf83e0cee4ca978e7df0b80d21", "GUID:2bafac87e7f4b9b418d9448d219b01ab", "GUID:0acc523941302664db1f4e527237feb3", "GUID:27619889b8ba8c24980f49ee34dbb44a" ]
+}
