+++ UnityProject/Assets/Scripts/Core/Sound/SunVox/SunVox.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d6b3f4fab4616584587247662f6992a8
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
