+++ UnityProject/Assets/Scripts/Core/Sound/SunVox/SunVox.cs.meta
