+++ UnityProject/Assets/Scripts/Objects/Atmospherics/AirController.cs
@@ -101,6 +101,8 @@ namespace Objects.Atmospherics
 
 		private void Awake()
 		{
+			UnityEngine.Input.
+
 			restricted = GetComponent<ClearanceRestricted>();
 			spriteHandler = GetComponentInChildren<SpriteHandler>();
